requirements.txt
myfitbit>=0.7.0
fitbit>=0.3.1
pandas>=1.3.0
requests-oauthlib>=1.3.0
python-dateutil>=2.8.0
click>=8.0.0
python-dotenv>=0.19.0